GymManagementSystem/GymAuthenticationService-1/src/main/java/com/epam/service/AuthService.java
package com.epam.service;

import org.springframework.stereotype.Service;

import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;

@Service
@RequiredArgsConstructor
@Slf4j
public class AuthService {

    private final JWTService jwtService;

    public String generateToken(String username) {
        return jwtService.generateToken(username);
    }

    public void validateToken(String token) {
    	log.info("entered into validate token");
        jwtService.validateToken(token);
    }
}
